requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0

# ── Tool System ────────────────────────────────────────────────────────────────
# DB tool (SQLAlchemy async)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0            # PostgreSQL async driver
aiomysql>=0.2.0            # MySQL async driver (optional)

# GCP tools (all optional — tools return ImportError if missing)
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
google-cloud-pubsub>=2.0.0
